Bug 573542 – blur plugin: bug in the first line
2009-03-12  Sven Neumann  <sven@gimp.org>

	Bug 573542 – blur plugin: bug in the first line

	* plug-ins/common/blur.c (blur_prepare_row): cast GimpPixelRgn.h
	to a signed integer.


svn path=/branches/gimp-2-6/; revision=28141

+++ plug-ins/common/blur.c
@@ -213,7 +213,7 @@ blur_prepare_row (GimpPixelRgn *pixel_rgn,
 {
   gint b;
 
-  y = CLAMP (y, 0, pixel_rgn->h - 1);
+  y = CLAMP (y, 0, (gint)pixel_rgn->h - 1);
 
   gimp_pixel_rgn_get_row (pixel_rgn, data, x, y, w);
